pyproject.toml
[project]
name = "dpl-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.12.0",
]
